producer/requirements.txt
kafka-python==2.3.1
numpy==2.4.4
pandas==3.0.2
py4j==0.10.9.9
python-dateutil==2.9.0.post0
six==1.17.0
